+++ sope-etl/src/main/scala/com/sope/etl/yaml/YamlFile.scala
@@ -1,8 +1,9 @@
-package com.sope.etl.transform.model
+package com.sope.etl.yaml
 
 import com.sope.etl.transform.Transformer
-import com.sope.etl.transform.YamlParserUtil._
 import com.sope.etl.transform.exception.YamlDataTransformException
+import com.sope.etl.transform.model.{TransformModel, TransformModelWithSourceTarget, TransformModelWithoutSourceTarget}
+import com.sope.etl.yaml.YamlParserUtil._
 import org.apache.spark.sql.{DataFrame, SQLContext}
 
 /**
@@ -10,7 +11,9 @@ import org.apache.spark.sql.{DataFrame, SQLContext}
   *
   * @author mbadgujar
   */
-class YamlFile[T <: TransformModel](yamlPath: String, substitutions: Option[Seq[Any]] = None, modelClass : Class[T]) {
+abstract class YamlFile[T <: TransformModel](yamlPath: String, substitutions: Option[Seq[Any]] = None, modelClass : Class[T]) {
+
+  protected val model: T = serialize
 
   private def updatePlaceHolders(): String = {
     substitutions.get
@@ -47,8 +50,6 @@ object YamlFile {
   case class IntermediateYaml(yamlPath: String, substitutions: Option[Seq[Any]] = None)
     extends YamlFile(yamlPath, substitutions, classOf[TransformModelWithoutSourceTarget]){
 
-    private val model = serialize
-
     /**
       * Perform transformation on provided dataframes.
       * The sources provided in YAML file should be equal and in-order to the provided dataframes
@@ -56,17 +57,15 @@ object YamlFile {
       * @return Transformed [[DataFrame]]
       */
     def getTransformedDFs(dataFrames: DataFrame*): Seq[(String, DataFrame)] = {
-      val transformModel = model.asInstanceOf[TransformModelWithoutSourceTarget]
-      if (transformModel.sources.size != dataFrames.size)
+      if (model.sources.size != dataFrames.size)
         throw new YamlDataTransformException("Invalid Dataframes provided or incorrect yaml config")
-      val sourceDFMap = transformModel.sources.zip(dataFrames).map { case (source, df) => (source, df.alias(source)) }
-      new Transformer(getYamlFileName, sourceDFMap.toMap, transformModel.transformations).transform
+      val sourceDFMap = model.sources.zip(dataFrames).map { case (source, df) => (source, df.alias(source)) }
+      new Transformer(getYamlFileName, sourceDFMap.toMap, model.transformations).transform
     }
   }
 
   case class End2EndYaml(yamlPath: String) extends YamlFile(yamlPath, None, classOf[TransformModelWithSourceTarget]){
 
-    private val model = serialize
     /**
       * Performs end to end transformations - Reading sources and writing transformation result to provided targets
       * The source yaml file should contains source and target information.
@@ -74,14 +73,12 @@ object YamlFile {
       * @param sqlContext Spark [[SQLContext]]
       */
     def performTransformations(sqlContext: SQLContext): Unit = {
-      val transformModel = model.asInstanceOf[TransformModelWithSourceTarget]
-      val sourceDFMap = transformModel.asInstanceOf[TransformModelWithSourceTarget].sources.map(source => source.getSourceName
+      val sourceDFMap = model.sources.map(source => source.getSourceName
         -> source.apply(sqlContext).alias(source.getSourceName)).toMap
-      val transformationResult = new Transformer(getYamlFileName, sourceDFMap, transformModel.transformations).transform.toMap
-      transformModel.targets.foreach(target => target(transformationResult(target.getInput)))
+      val transformationResult = new Transformer(getYamlFileName, sourceDFMap, model.transformations).transform.toMap
+      model.targets.foreach(target => target(transformationResult(target.getInput)))
     }
   }
-
 }
